Rename user administration components directory
